Move to another file, fix issue_category
